refactor(orchestrator): extract MCP-call telemetry from executor_agent_loop
Last sub-PR of the executor_agent_loop.go split. recordMCPCallTelemetry
and its mcpToolInputForLog helper had become the only large
non-dispatched concern left inline — they're called from dispatchMCPToolCall
(executor_agent_loop_tools.go) and own the per-call metric / event /
input-capture surface. Move them to their own file so the executor's
main file shrinks to the Execute loop + assistant-event emitters +
agentToolDefinitions + step/artifact builders.

  executor_agent_loop.go:            1610 → 1541 LOC
  executor_agent_loop_telemetry.go:  new, 78 LOC

What moves to executor_agent_loop_telemetry.go

  - (e *AgentLoopExecutor).recordMCPCallTelemetry: emits the
    per-dispatch telemetry.Metrics.RecordMCP + trace.Record event
    pair around an MCP tool call. Used by all four exit paths in
    dispatchMCPToolCall (blocked / dispatched / tool_error /
    failed) so they share one telemetry surface.
  - mcpToolInputForLog: builds the trace-friendly input-log map
    for an MCP call (drops oversize args, captures arg keys when
    a value is non-scalar). Used by dispatchMCPToolCall too.

What stays in executor_agent_loop.go

  Construction (NewAgentLoopExecutor / SetMetrics /
  SetMCPHostFactory), the main Execute loop, the assistant-event
  emitters (emitAgentTurnStarted / emitAssistantMessageEvents /
  emitAssistantFinalAnswer), the conversation-artifact upsert
  (upsertConversationArtifact), step/artifact builders, and the
  agentToolDefinitions catalog. The four other concern files
  (executor_agent_loop_chat.go, _tools.go, _http.go, _telemetry.go)
  now own their respective surfaces.

The receiver type stays *AgentLoopExecutor; no exported symbol
moves; tests untouched. executor_agent_loop_telemetry.go takes its
own context/encoding-json/strings + internal/telemetry + pkg/types.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
